refactor: GameManger에 종속적인 ExecuteStatus를 Nested Enum으로 수정
